pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "minifold"
version = "0.1.0"
requires-python = ">=3.9"
description = "MiniFold"
readme = "README.md"
dependencies = [
    "numpy",
    "scipy",
    "pytorch_lightning",
    "torch>=2.3.0",
    "fair-esm",
    "biopython==1.81",
    "edit_distance",
    "pandas",
    "modelcif",
    "dm-tree",
    "psutil",
    "click",
    "ml_collections",
    "einops",
]

[tool.setuptools]
packages = ["minifold"]

[tool.setuptools.package-data]
minifold = ["utils/stereo_chemical_props.txt"]
